requirements.txt
# requirements.txt
tensorflow
numpy
mido
pretty_midi
tqdm
pyfluidsynth
datasets